requirements.txt
requests
pandas
fastapi
uvicorn[standard]
pydantic[email]
PyJWT
bcrypt